requirements.txt
streamlit
easyocr
pillow
numpy
fuzzywuzzy